Move extra helpers. Cleanup
