+++ waspc/data/Generator/templates/server/src/auth/providers/oauth/user.ts
@@ -12,8 +12,7 @@ import { type {= authEntityUpper =} } from 'wasp/entities'
 import { prisma } from 'wasp/server'
 import { type UserSignupFields, type ProviderConfig } from 'wasp/auth/providers/types'
 import { type OAuthParams } from 'wasp/server/auth'
-import { getRedirectUriForOneTimeCode } from './redirect'
-import { tokenStore } from './oneTimeCode'
+import { getRedirectUriForOneTimeCode, tokenStore } from 'wasp/server/oauth'
 import {
   onBeforeSignupHook,
   onAfterSignupHook,
